+++ .gitignore
@@ -1,0 +1,4 @@
+*.xml
+*.txt
+idea/
+*.class
