+++ lab2/input.txt
@@ -1,5 +1,5 @@
 #Пример (должно вывести 2):
-DEFINE a 4
+DEFINE a -4
 PUSH a
 SQRT
 PRINT
