--- lab2/src/config.properties
@@ -1,1 +1,0 @@
-OPERATIONS_CLASS_PATH=resources.Operations.
