--- lab2/src/logs/lab2log.log
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<!DOCTYPE log SYSTEM "logger.dtd">
-<log>
-<record>
-  <date>2022-03-07T01:57:39.690436800Z</date>
-  <millis>1646618259690</millis>
-  <nanos>436800</nanos>
-  <sequence>0</sequence>
-  <logger>Main</logger>
-  <level>INFO</level>
-  <class>Main</class>
-  <method>main</method>
-  <thread>1</thread>
-  <message>file reading start</message>
-</record>
-<record>
-  <date>2022-03-07T01:57:39.778441800Z</date>
-  <millis>1646618259778</millis>
-  <nanos>441800</nanos>
-  <sequence>1</sequence>
-  <logger>Main</logger>
-  <level>WARNING</level>
-  <class>Main</class>
-  <method>main</method>
-  <thread>1</thread>
-  <message>warning </message>
-</record>
-<record>
-  <date>2022-03-07T01:57:39.780441900Z</date>
-  <millis>1646618259780</millis>
-  <nanos>441900</nanos>
-  <sequence>2</sequence>
-  <logger>Main</logger>
-  <level>INFO</level>
-  <class>Main</class>
-  <method>main</method>
-  <thread>1</thread>
-  <message>reading was successful</message>
-</record>
-</log>
