+++ lab2/src/main/Exceptions/WrongArgumentsAmountException.java
@@ -2,6 +2,6 @@ package Exceptions;
 
 public class WrongArgumentsAmountException extends Exception {
     public WrongArgumentsAmountException(String operationName, int argsAmount){
-        super("in operation " + operationName + "expected "+ argsAmount + " arguments");
+        super("in operation " + operationName + " expected "+ argsAmount + " arguments");
     }
 }
