+++ lab2/src/main/OperationsAndFactory/Operations/DivOperation.java
@@ -1,21 +1,21 @@
-package resources.Operations;
+package OperationsAndFactory.Operations;
 
 import Exceptions.WrongArgumentsAmountException;
-import resources.Operation;
+import OperationsAndFactory.Operation;
 
 import java.util.HashMap;
 import java.util.Stack;
 
-public class SumOperation implements Operation {
+public class DivOperation implements Operation {
     private int argsAmount = 0;
     @Override
     public void Execute(Stack<Double> stack, HashMap<String, String> defineMap, String[] args)
             throws WrongArgumentsAmountException {
         if(args.length != argsAmount){
-            throw new WrongArgumentsAmountException("DEFINE", argsAmount);
+            throw new WrongArgumentsAmountException("DIV", argsAmount);
         }
         double a = stack.pop();
         double b = stack.pop();
-        stack.add(a+b);
+        stack.add(b / a);
     }
 }
