+++ lab2/src/main/OperationsAndFactory/Operations/PrintOperation.java
@@ -1,21 +1,19 @@
-package resources.Operations;
+package OperationsAndFactory.Operations;
 
 import Exceptions.WrongArgumentsAmountException;
-import resources.Operation;
+import OperationsAndFactory.Operation;
 
 import java.util.HashMap;
 import java.util.Stack;
 
-public class DivOperation implements Operation {
+public class PrintOperation implements Operation {
     private int argsAmount = 0;
     @Override
     public void Execute(Stack<Double> stack, HashMap<String, String> defineMap, String[] args)
             throws WrongArgumentsAmountException {
         if(args.length != argsAmount){
-            throw new WrongArgumentsAmountException("DEFINE", argsAmount);
+            throw new WrongArgumentsAmountException("PRINT", argsAmount);
         }
-        double a = stack.pop();
-        double b = stack.pop();
-        stack.add(b / a);
+        System.out.println(stack.peek());
     }
 }
