+++ lab2/src/main/OperationsAndFactory/Operations/PushOperation.java
@@ -1,19 +1,19 @@
-package resources.Operations;
+package OperationsAndFactory.Operations;
 
 import Exceptions.WrongArgumentsAmountException;
-import resources.Operation;
+import OperationsAndFactory.Operation;
 
 import java.util.HashMap;
 import java.util.Stack;
 
-public class PrintOperation implements Operation {
-    private int argsAmount = 0;
+public class PushOperation implements Operation {
+    private int argsAmount = 1;
     @Override
     public void Execute(Stack<Double> stack, HashMap<String, String> defineMap, String[] args)
             throws WrongArgumentsAmountException {
         if(args.length != argsAmount){
-            throw new WrongArgumentsAmountException("DEFINE", argsAmount);
+            throw new WrongArgumentsAmountException("PUSH", argsAmount);
         }
-        System.out.println(stack.peek());
+        stack.push(Double.parseDouble(args[0]));
     }
 }
