+++ lab2/src/main/OperationsAndFactory/Operations/SqrtOperation.java
@@ -1,19 +1,22 @@
-package resources.Operations;
+package OperationsAndFactory.Operations;
 
 import Exceptions.WrongArgumentsAmountException;
-import resources.Operation;
+import OperationsAndFactory.Operation;
 
 import java.util.HashMap;
 import java.util.Stack;
 
-public class PushOperation implements Operation {
-    private int argsAmount = 1;
+public class SqrtOperation implements Operation {
+    private int argsAmount = 0;
     @Override
     public void Execute(Stack<Double> stack, HashMap<String, String> defineMap, String[] args)
             throws WrongArgumentsAmountException {
         if(args.length != argsAmount){
-            throw new WrongArgumentsAmountException("DEFINE", argsAmount);
+            throw new WrongArgumentsAmountException("SQRT", argsAmount);
         }
-        stack.push(Double.parseDouble(args[0]));
+        double val  = stack.pop();
+        stack.push(Math.sqrt(val));
+
     }
+
 }
