+++ lab2/src/main/OperationsAndFactory/Operations/SubOperation.java
@@ -1,7 +1,7 @@
-package resources.Operations;
+package OperationsAndFactory.Operations;
 
 import Exceptions.WrongArgumentsAmountException;
-import resources.Operation;
+import OperationsAndFactory.Operation;
 
 import java.util.HashMap;
 import java.util.Stack;
@@ -12,7 +12,7 @@ public class SubOperation implements Operation {
     public void Execute(Stack<Double> stack, HashMap<String, String> defineMap, String[] args)
             throws WrongArgumentsAmountException {
         if(args.length != argsAmount){
-            throw new WrongArgumentsAmountException("DEFINE", argsAmount);
+            throw new WrongArgumentsAmountException("SUB", argsAmount);
         }
         double a = stack.pop();
         double b = stack.pop();
