+++ lab2/src/main/OperationsAndFactory/Operations/SumOperation.java
@@ -1,21 +1,21 @@
-package resources.Operations;
+package OperationsAndFactory.Operations;
 
 import Exceptions.WrongArgumentsAmountException;
-import resources.Operation;
+import OperationsAndFactory.Operation;
 
 import java.util.HashMap;
 import java.util.Stack;
 
-public class MulOperation implements Operation {
+public class SumOperation implements Operation {
     private int argsAmount = 0;
     @Override
     public void Execute(Stack<Double> stack, HashMap<String, String> defineMap, String[] args)
             throws WrongArgumentsAmountException {
         if(args.length != argsAmount){
-            throw new WrongArgumentsAmountException("DEFINE", argsAmount);
+            throw new WrongArgumentsAmountException("SUM", argsAmount);
         }
         double a = stack.pop();
         double b = stack.pop();
-        stack.add(b * a);
+        stack.add(a+b);
     }
 }
