+++ lab2/src/main/logging/logs.properties
@@ -1,6 +1,6 @@
 handlers= java.util.logging.FileHandler, java.util.logging.ConsoleHandler
 .level=INFO
-java.util.logging.FileHandler.pattern = C:/Users/Alexander/IdeaProjects/lab2/src/logs/lab2log.log
+java.util.logging.FileHandler.pattern = C:/Users/Alexander/IdeaProjects/lab2/src/main/logs/lab2log.log
 java.util.logging.FileHandler.limit = 50000
 java.util.logging.FileHandler.count = 1
 java.util.logging.FileHandler.formatter = java.util.logging.XMLFormatter
