+++ lab2/src/main/logs/lab2log.log
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<!DOCTYPE log SYSTEM "logger.dtd">
+<log>
+<record>
+  <date>2022-03-18T14:32:50.774793100Z</date>
+  <millis>1647613970774</millis>
+  <nanos>793100</nanos>
+  <sequence>0</sequence>
+  <logger>Main</logger>
+  <level>INFO</level>
+  <class>Main</class>
+  <method>main</method>
+  <thread>1</thread>
+  <message>start of reading </message>
+</record>
+<record>
+  <date>2022-03-18T14:33:27.002865200Z</date>
+  <millis>1647614007002</millis>
+  <nanos>865200</nanos>
+  <sequence>1</sequence>
+  <logger>Calculator</logger>
+  <level>WARNING</level>
+  <class>Calculator</class>
+  <method>calculate</method>
+  <thread>1</thread>
+  <message>calculating error: in operation SQRT expected 0 arguments</message>
+</record>
+</log>
