+++ lab2/src/main/operations.properties
